Moved sequence adaptation loading code into SequenceEditor to include it in the sequence preview
